refactor: apply FSD structure
