+++ src/app/index.tsx
@@ -1,12 +1,14 @@
 import React, { Component } from 'react';
-import { Search } from './components/search';
-import { ErrorBoundary } from './components/errorBoundary';
+import { Search } from '../widgets/search';
+import { ErrorBoundary } from './errorBoundary';
+import { TestError } from '../entities/testError';
 
 export class App extends Component {
   render() {
     return (
       <ErrorBoundary>
         <Search />
+        <TestError />
       </ErrorBoundary>
     );
   }
