+++ src/entities/input.tsx
