+++ src/shared/api.ts
@@ -1,4 +1,4 @@
-import { UserData } from '../types';
+import { UserData } from './types';
 
 export class Api {
   static instance: Api;
